Moved getSettings call to hide Data source panel for cluster metrics monitors when remote monitoring is disabled.
