+++ public/pages/CreateMonitor/containers/DataSource/DataSource.js
@@ -33,7 +33,7 @@ class DataSource extends Component {
   }
 
   render() {
-    const { isMinimal } = this.props;
+    const { isMinimal, remoteMonitoringEnabled } = this.props;
     const { monitor_type, searchType } = this.props.values;
     const displayTimeField =
       searchType === SEARCH_TYPE.GRAPH &&
@@ -41,7 +41,12 @@ class DataSource extends Component {
       monitor_type !== MONITOR_TYPE.CLUSTER_METRICS;
     const monitorIndexDisplay = (
       <>
-        <MonitorIndex httpClient={this.props.httpClient} monitorType={monitor_type} />
+        <MonitorIndex
+          httpClient={this.props.httpClient}
+          monitorType={monitor_type}
+          remoteMonitoringEnabled={remoteMonitoringEnabled}
+        />
+
         {displayTimeField && (
           <>
             <EuiSpacer />
