+++ public/pages/CreateMonitor/containers/DefineMonitor/DefineMonitor.js
@@ -39,6 +39,7 @@ import ConfigureDocumentLevelQueries from '../../components/DocumentLevelMonitor
 import FindingsDashboard from '../../../Dashboard/containers/FindingsDashboard';
 import { validDocLevelGraphQueries } from '../../components/DocumentLevelMonitorQueries/utils/helpers';
 import { validateWhereFilters } from '../../components/MonitorExpressions/expressions/utils/whereHelpers';
+import { REMOTE_MONITORING_ENABLED_SETTING_PATH } from '../../components/CrossClusterConfigurations/components/ExperimentalBanner';
 
 function renderEmptyMessage(message) {
   return (
@@ -74,6 +75,7 @@ class DefineMonitor extends Component {
       plugins: [],
       loadingResponse: false,
       PanelComponent: props.flyoutMode ? ({ children }) => <>{children}</> : ContentPanel,
+      remoteMonitoringEnabled: false,
     };
 
     this.renderGraph = this.renderGraph.bind(this);
@@ -88,6 +90,7 @@ class DefineMonitor extends Component {
     this.getPlugins = this.getPlugins.bind(this);
     this.getSupportedApiList = this.getSupportedApiList.bind(this);
     this.showPluginWarning = this.showPluginWarning.bind(this);
+    this.getSettings = this.getSettings.bind(this);
   }
 
   componentDidMount() {
@@ -96,6 +99,7 @@ class DefineMonitor extends Component {
     const isGraph = searchType === SEARCH_TYPE.GRAPH;
     const hasIndices = !!index.length;
     const hasTimeField = !!timeField;
+    this.getSettings();
     if (isGraph && hasIndices) {
       this.onQueryMappings();
       if (hasTimeField || !this.requiresTimeField()) this.onRunQuery();
@@ -172,6 +176,34 @@ class DefineMonitor extends Component {
     if (prevSearchType !== searchType || prevMonitorType !== monitor_type || groupByCleared) {
       this.resetResponse();
       if (searchType === SEARCH_TYPE.CLUSTER_METRICS) this.getSupportedApiList();
+    }
+  }
+
+  async getSettings() {
+    try {
+      const { httpClient } = this.props;
+      const response = await httpClient.get('../api/alerting/_settings');
+      if (response.ok) {
+        const { defaults, transient, persistent } = response.resp;
+        let remoteMonitoringEnabled = _.get(
+          // If present, take the 'transient' setting.
+          transient,
+          REMOTE_MONITORING_ENABLED_SETTING_PATH,
+          // Else take the 'persistent' setting.
+          _.get(
+            persistent,
+            REMOTE_MONITORING_ENABLED_SETTING_PATH,
+            // Else take the 'default' setting.
+            _.get(defaults, REMOTE_MONITORING_ENABLED_SETTING_PATH, false)
+          )
+        );
+        // Boolean settings are returned as strings (e.g., `"true"`, and `"false"`). Constructing boolean value from the string.
+        if (typeof remoteMonitoringEnabled === 'string')
+          remoteMonitoringEnabled = JSON.parse(remoteMonitoringEnabled);
+        this.setState({ remoteMonitoringEnabled: remoteMonitoringEnabled });
+      }
+    } catch (e) {
+      console.log('Error while retrieving settings', e);
     }
   }
 
@@ -614,19 +646,27 @@ class DefineMonitor extends Component {
   }
 
   render() {
-    const { values, errors, httpClient, detectorId, notifications, isDarkMode, flyoutMode } =
-      this.props;
-    const { dataTypes, PanelComponent } = this.state;
+    const {
+      values,
+      values: { monitor_type },
+      errors,
+      httpClient,
+      detectorId,
+      notifications,
+      isDarkMode,
+      flyoutMode,
+    } = this.props;
+    const { dataTypes, PanelComponent, remoteMonitoringEnabled } = this.state;
     const monitorContent = this.getMonitorContent();
     const { searchType } = this.props.values;
-    const isGraphOrQuery =
+    const displayDataSourcePanel =
       searchType === SEARCH_TYPE.GRAPH ||
       searchType === SEARCH_TYPE.QUERY ||
-      this.props.values.monitor_type === MONITOR_TYPE.CLUSTER_METRICS;
+      (remoteMonitoringEnabled && monitor_type === MONITOR_TYPE.CLUSTER_METRICS);
 
     return (
       <div>
-        {!flyoutMode && isGraphOrQuery && (
+        {!flyoutMode && displayDataSourcePanel && (
           <div>
             <DataSource
               values={values}
@@ -636,6 +676,7 @@ class DefineMonitor extends Component {
               detectorId={detectorId}
               notifications={notifications}
               isDarkMode={isDarkMode}
+              remoteMonitoringEnabled={remoteMonitoringEnabled}
             />
             <EuiSpacer />
           </div>
